requirements.txt
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
stripe==14.0.1
